Update Coupon CRUD - Handle Coupons App

+++ database/migrations/2024_10_20_075918_create_packages_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name_ar');
             $table->string('name_en');
-            $table->integer('views');
+            $table->unsignedBigInteger('views');
             $table->decimal('price');
             $table->string('color');
             $table->boolean('is_active')->default(1);
